feat #459: Move Shop to Client routes
